split metal/f0; fix legacy/lab encodings; double-precision and range/value scaling

+++ PixelGraph.Common/Material/MaterialAlbedoProperties.cs
@@ -7,15 +7,15 @@ namespace PixelGraph.Common.Material
         public string Texture {get; set;}
 
         public ResourcePackAlbedoRedChannelProperties InputRed {get; set;}
-        public byte? ValueRed {get; set;}
+        public decimal? ValueRed {get; set;}
         public decimal? ScaleRed {get; set;}
 
         public ResourcePackAlbedoGreenChannelProperties InputGreen {get; set;}
-        public byte? ValueGreen {get; set;}
+        public decimal? ValueGreen {get; set;}
         public decimal? ScaleGreen {get; set;}
 
         public ResourcePackAlbedoBlueChannelProperties InputBlue {get; set;}
-        public byte? ValueBlue {get; set;}
+        public decimal? ValueBlue {get; set;}
         public decimal? ScaleBlue {get; set;}
     }
 }
