+++ PixelGraph.Common/Material/MaterialHeightProperties.cs
@@ -6,8 +6,8 @@ namespace PixelGraph.Common.Material
     {
         public ResourcePackHeightChannelProperties Input {get; set;}
         public string Texture {get; set;}
-        public byte? Value {get; set;}
-        public int? Shift {get; set;}
+        public decimal? Value {get; set;}
+        public decimal? Shift {get; set;}
         public decimal? Scale {get; set;}
         public int? EdgeFadeSizeX {get; set;}
         public int? EdgeFadeSizeY {get; set;}
